requirements.txt
asgiref==3.10.0
Django==5.2.7
pillow==11.3.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2

# Security & Cache
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
## The following requirements were added by pip freeze:
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.2
deprecation==2.1.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
postgrest==2.22.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
realtime==2.22.0
requests==2.32.5
sniffio==1.3.1
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
## The following requirements were added by pip freeze:
colorama==0.4.6
distro==1.9.0
git-filter-repo==2.47.0
groq==0.32.0
jiter==0.11.0
openai==1.54.3
python-dotenv==1.1.1
tqdm==4.67.1